[MRM-747] Archiva should prevent re-deployment of released or non-snapshot versioned artifacts o block re-deployment of released artifacts to repository of configured using web upload form o added unit test
git-svn-id: https://svn.apache.org/repos/asf/archiva/trunk@883057 13f79535-47bb-0310-9956-ffa450edef68

+++ archiva-modules/archiva-web/archiva-webapp/src/main/java/org/apache/maven/archiva/web/action/UploadAction.java
@@ -351,8 +351,17 @@ public class UploadAction
             
             try
             {   
-                copyFile( artifactFile, targetPath, filename, fixChecksums );
-                queueRepositoryTask( repository.getId(), repository.toFile( artifactReference ) );
+                File targetFile = new File( targetPath, filename );
+                if( targetFile.exists() && !VersionUtil.isSnapshot( version ) && repoConfig.isBlockRedeployments() )
+                {
+                    addActionError( "Overwriting released artifacts in repository '" + repoConfig.getId() + "' is not allowed." );
+                    return ERROR;
+                }
+                else
+                {
+                    copyFile( artifactFile, targetPath, filename, fixChecksums );
+                    queueRepositoryTask( repository.getId(), repository.toFile( artifactReference ) );                    
+                }
             }
             catch ( IOException ie )
             {
